is operator

+++ petrov.vadim/S2/foo.hpp
@@ -7,7 +7,7 @@
 namespace petrov
 {
   bool isOperator(const std::string& token);
-  int getPriority(const std::string& op);
+  size_t getPriority(const std::string& op);
   size_t reverseNumber(size_t num);
   bool isNumber(const std::string& token);
 }
